+++ src/providers/pronouns/alejo/PronounsAlejoApi.hpp
@@ -2,15 +2,20 @@
 
 #include "providers/pronouns/UserPronouns.hpp"
 
+#include <QJsonObject>
+#include <QString>
+
 #include <mutex>
 #include <optional>
+#include <shared_mutex>
 
-namespace chatterino::Pronouns {
+namespace chatterino::pronouns {
 
 class AlejoApi
 {
 private:
-    std::mutex mutex;
+    std::shared_mutex mutex;
+    /** A map from alejo.io ids to human readable representation like theythem -> they/them, other -> other. */
     std::optional<std::unordered_map<QString, QString>> pronounsFromId =
         std::nullopt;
     UserPronouns parse(QJsonObject);
@@ -20,8 +25,11 @@ private:
 
 public:
     explicit AlejoApi();
+    /** Fetches pronouns from the alejo.io API for a username and calls onDone when done.
+        onDone can be invoked from any thread. The argument is std::nullopt if and only if the request failed.
+     */
     void fetch(const QString &user,
                std::function<void(std::optional<UserPronouns>)> onDone);
 };
 
-}  // namespace chatterino::Pronouns
+}  // namespace chatterino::pronouns